requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.13,<4.0
django-cors-headers>=3.10,<4.0
celery>=5.2,<6.0
redis>=4.0,<5.0
psycopg2-binary>=2.9,<3.0
Pillow>=9.0,<10.0
python-dotenv==1.0.1